Mv BaseURL Option generator to tests
